Ensure a consistent order for class forward-declarations
The forward-declarations would show up in a different order each time
the application was restarted and the model regenerated. This was
actually due to a feature of QSet, which randomizes the hashes of its
values with a seed.

Resolved this by converting the set to a list and sorting it before
writing out the list.